+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcCat/DmcCatAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcCat
+{
+	[KrangSettings("DmcCat", typeof(DmcCatAdapter))]
+	public sealed class DmcCatAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}