+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcCatDsp/DmcCatDspAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcCatDsp
+{
+	[KrangSettings("DmcCatDsp", typeof(DmcCatDspAdapter))]
+	public sealed class DmcCatDspAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}