+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcDvi/DmcDviAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcDvi
+{
+	[KrangSettings("DmcDvi", typeof(DmcDviAdapter))]
+	public sealed class DmcDviAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}