+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcF/DmcFAdapterRoutingControl.cs
@@ -1,0 +1,19 @@
+﻿#if SIMPLSHARP
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcF
+{
+	public sealed class DmcFAdapterRoutingControl : AbstractInputCardAdapterRoutingControl<DmcFAdapter>
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		public DmcFAdapterRoutingControl(DmcFAdapter parent, int id)
+			: base(parent, id)
+		{
+		}
+	}
+}
+
+#endif