+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcF/DmcFAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcF
+{
+	[KrangSettings("DmcF", typeof(DmcFAdapter))]
+	public sealed class DmcFAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}