+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcFDsp/DmcFDspAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcFDsp
+{
+	[KrangSettings("DmcFDsp", typeof(DmcFDspAdapter))]
+	public sealed class DmcFDspAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}