+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcS/DmcSAdapterRoutingControl.cs
@@ -1,0 +1,19 @@
+﻿#if SIMPLSHARP
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcS
+{
+	public sealed class DmcSAdapterRoutingControl : AbstractInputCardAdapterRoutingControl<DmcSAdapter>
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		public DmcSAdapterRoutingControl(DmcSAdapter parent, int id)
+			: base(parent, id)
+		{
+		}
+	}
+}
+
+#endif