+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcS/DmcSAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcS
+{
+	[KrangSettings("DmcS", typeof(DmcSAdapter))]
+	public sealed class DmcSAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}