+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcSDsp/DmcSDspAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcSDsp
+{
+	[KrangSettings("DmcSDsp", typeof(DmcSDspAdapter))]
+	public sealed class DmcSDspAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}