+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcSdi/DmcSdiAdapterRoutingControl.cs
@@ -1,0 +1,19 @@
+﻿#if SIMPLSHARP
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcSdi
+{
+	public sealed class DmcSdiAdapterRoutingControl : AbstractInputCardAdapterRoutingControl<DmcSdiAdapter>
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		public DmcSdiAdapterRoutingControl(DmcSdiAdapter parent, int id)
+			: base(parent, id)
+		{
+		}
+	}
+}
+
+#endif