+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcSdi/DmcSdiAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcSdi
+{
+	[KrangSettings("DmcSdi", typeof(DmcSdiAdapter))]
+	public sealed class DmcSdiAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}