+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/DmcStr/DmcStrAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.DmcStr
+{
+	[KrangSettings("DmcStr", typeof(DmcStrAdapter))]
+	public sealed class DmcStrAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}